requirements.txt
matplotlib==3.10.0
mysql-connector-python==9.2.0
openai==0.28.0


pandas==2.2.3
streamlit==1.41.1
